Add resolveDescriptorGet() helper function
There are already a few places that can use this, and I have more
additions in upcoming caching diffs.

For the now-unused `Frame*` parameter in `lookupMethod()`: I decided to leave
this rather than doing a big codemod to eliminate it from all callers now,
since it's going to get a big signature change with T44238481 anyway.

Based on Facebook D15362535