Introduce StrArray type
Introduces a `MutableBytes` backed `StrArray` to replace our usage of `ByteArray` as a string builder.

Since the only way to add to a `StrArray` is by appending a codepoint or an already well-formed `Str`, it maintains an invariant that at all points, a `StrArray` contains valid UTF-8.

This diff adds the API that `StrArray` needs to become the codec system's string builder

Based on Facebook D15341523