Shortcut condition evaluation if value is already a boolean
Avoid calling into `Interpreter::isTrue()` when the value is already a boolean.

This improves performance as it avoids creating `HandleScope`s in a hot part of the interpreter. This also changes the signature of `isTrue` so we do not even need the HandleScope in the `else` part of the calling functions.

While on it also fix exception unwinding in the jump opcodes.

Based on Facebook D15376457