Add _strarray builtins
Add `strarray-builtins*` to ensure that the type is accessible from managed code.

Codecs needs `_strarray` to be able to be turned into a `str` and to be able to append `str`s to it, so those are the only dunder functions I've added in this diff.

Since `__repr__`s are supposed to be used for debugging purposes, I decided to have its repr function return the fact that it's a `_strarray`, similar to how `bytearray` returns `bytearray(b'')`.

Based on Facebook D15358257