Remove expensive and unnecessary DCHECK in `RawArray::length()`
I ran a DebugOpt build of `python-tests` under `perf`, and we were
spending about 20% of CPU time in `isHeapObjectWithLayout()` on behalf of this
`DCHECK` (mostly from the range check in `RawTuple::at()`, which is worth
keeping). It's burning cycles to check a static property of the way we
structure our code: as long as we don't let random crap inherit from
`RawArray`, there's nothing to verify here.

Based on Facebook D15402470