Optimize HandleScope and Handles
There are a few changes here:

1) Rather than each `Thread` having a `Vector<HandleScope*>`, and each
   `HandleScope` being the root of a linked list of `Handle`s, keep a single
   linked list of `Handle`s per `Thread`, in the newly-trimmed-down
   `Handles`. `HandleScope` is still around in case we want to make the
   distinction useful again, and in a few experiments the compiler was pretty
   good at optimizing it away.
2) Get rid of all of the `friend` declarations from the involved classes and
   make the necessary bits public. The API surface area between `Handles`,
   `Handle<T>`, and `HandleScope` is simpler now, so I think this is net nicer.
3) Don't store a `Thread*` in either `HandleScope` or `Handle`. It was only
   used for a debug-only type check, and using `Thread::current()` there is
   fine (this diff is still a 9% CPU time win for DebugOpt builds).

The generated code for constructing and destructing `Handle`s still leaves a
lot to be desired, but short of fundamentally changing how we construct them
(in a big codemod), there's not much else that can be done right now. Plus,
`Handle`s should become less relevant over time, so this isn't a great area to
sink much more time.

The changes in `handles-test.cpp` are because we no longer verify that
`Handle`s are created in the topmost scope (since it doesn't matter), and there
was an unused `Handles` variable that the compiler was able to warn about now.

Based on Facebook D15460833