Rewrite traceback printer to access function instead of code object in frame
This changes the traceback printer to not use `Frame::code()` any more in favor of `Frame::function()`.

This has some slight effects like us printing the name of the function now instead of the name of the code object and I slightly changed the format of printing native frames.

This change is in preparation of removing the code field from `Frame`.

Based on Facebook D15451358