Remove code field from frame
Now that `Frame::function()` is guaranteed to work in all (relevant) situations, we can stop putting the code object into the frame, and simply grab it from the function object in the few cases that still need it.

This is a small perf win as we have to write less when setting up a new frame.

- Remove `Frame::setCode()`, keep `Frame::code()` but make it call `Frame::function().code()`.
- `Thread::makeNativeFrame()` has nothing left to do, so I removed it.
- I had to add `RawHeapFrame::function()` because accessing any locals in a HeapFrame fails without special handling.

Based on Facebook D15451979