Fix TryBlock level bug in Interpreter::doWithCleanupStart()
`TryBlock::setLevel()` looks like it was only meant to be used while
initializing a zero-filled object, but I was using it like a normal setter in
D13995720. `TryBlock` is just a wrapper around a `uword`, so rework the
initialization path to make this mistake impossible, and construct a new
`TryBlock` in `doWithCleanupStart()` instead.

Based on Facebook D15539758