Avoid empty entry check from attribute cache lookup
SmallInt is always different from None so we can just keep searching till the end of the cache line without checking if a cache entry is filled or not. Assuming cache hit ratio is high, dropping the None check can make cache lookup faster.

This diff saves 1.12% of instructions to run Richards.

```
{
  "benchmark_name": "richards.py",
  "instructions": 4014051117,
  "interpreter_name": "./opt/python-control",
  "run_time_secs": 0.6467920867726207,
  "version": "n/a"
}
{
  "benchmark_name": "richards.py",
  "instructions": 4059577429,
  "interpreter_name": "./opt/python-base",
  "run_time_secs": 0.6364405453205109,
  "version": "n/a"
}
```

Based on Facebook D15548400