Improve Python test failure messages
We don't have proper traceback support yet, so when a Python unittest
fails, nothing useful is printed. This makes debugging test failures an
enormous pain, and this diff is a temporary fix to reduce that pain until we do
support normal tracebacks.

The `Thread::raiseWithType()` refactor is meant to be permanent; my upcoming
diff for T40269344 also needs to hook into raising, whether it was from C++ or
managed code.

I also fixed the name of the `__traceback__` attribute on `BaseException`, and
fixed a "caught" vs. "pending" typo in `genDoThrowRaise()`, which were causing
issues with testing this.

If anyone's worried about the `HandleScope` or extra branch in
`raiseWithType()`, this diff appears to be a small but consistent win for some
reason:
```
./pyro.baseline -> ./pyro.after
  cycles        :  -2.6% to  -0.8%, mean  -1.7%
  instructions  :  -0.3% to  +0.2%, mean  -0.1% ✗
  branch-misses :  -3.7% to  -0.9%, mean  -2.3%
  branches      :  -0.1% to  +0.1%, mean  -0.0% ✗
```

Based on Facebook D15539962