Move newEmptyCode() back to test-utils.h
The pattern of creating an "empty" code object and then changing a couple fields in it makes it hard to do things like `DCHECK` for consistency or pre-computing derived values. I would like to discourage this programming style at least in the runtime by removing the `newEmptyCode()` function and forcing people to always use `Runtime::newCode()`.

This diff keeps `newEmptyCode()` as a testing-only function for now.

Based on Facebook D15549224