Implement PyUnicode_IS_ASCII
Summary: In CPython the `PyUnicode` tracks information about its chars and is able to resolve this in O(1) time. Given that this is not `PyUnicode_IS_ASCII` is implemented in O(n) time. The `Str` implementation does not currently track information about the chars so we can't resolve it in O(1) time.

Based on Facebook D15550096