Enable choosing a subset of benchmarks
Summary: This allows the benchmark.py to select a subset of benchmarks to run. This removes the need to pass in the benchmarks path as that can be automatically determined. Instead, we now need to pass in the benchmarks that are intended to be run. By default, it will run all of them.

Based on Facebook D15580332