Fix Code::totalVars()
`Code::totalVars()` is intended to capture the amount of locals that are not arguments (those are reported in `totalArgs()`).

The calculated value of `Code::totalVars()` was too high as it didn't subtract the arguments from `nlocals`.

While on it add some `DCHECK` to verify sensible `nlocals` values when creating a code object.

Based on Facebook D15586337