Implement _PyBytesWriter API
Implemented the _PyBytesWriter struct and its 5 associated C-API functions.

Some of the struct fields (allocated, min_size, overallocate, and use_bytearray) are accessed elsewhere in CPython, so I left them in the Pyro struct.

I made allocation overallocate when use_bytearray is set to mimic the built-in overallocation of ByteArray objects in CPython.

Based on Facebook D15577474