Make ascii_decode use _strarray
Replace `ascii_decode`'s usage of `bytearray` with `strarray` and remove `_bytearray_to_string`, since it's only purpose was to be used in `ascii_decode`.

Also adds a `strArrayAddASCII` function to the runtime to fast-path the common use-case of just adding ASCII bytes.

Part of the effort to stop abusing LargeBytes.

Based on Facebook D15395169