Ignore exception regexes in unittest
Disable use of `re.compile()` in unittest that doesn't work properly in pyro yet.

This allows to use `assertRaisesRegex()`. We will not check anything in the pyro version yet, but at least we can write the tests and have the checks performed with cpython until we have a working `re` module.

Based on Facebook D15605681