Do not include implicit globals into frame of optimized functions
Remove the implicit global field from the frame.

Implicit globals are only used for functions without the `CO_OPTIMIZED` and `CO_NEWLOCALS` flags set. This only happens for module bodies (and other forms of `exec()`/`eval()`) and class bodies, which both use specialized APIs (`Thread::exec()` and `Thread::runClassFunction()`). So instead of having a field in the frame that is always reserved, I push the implicit globals next to the function in `exec()` and `runClassFunction()` and modify `Frame::implicitGlobals()` to look for it there. In all other calls no space will be occupied.

Based on Facebook D15568975