Cache global vars with cache invalidation
This diff implements global variable caching and its invalidation as proposed in https://fb.quip.com/2ToaAzSiaYmA.

Under this change LOAD_GLOBAL / STORE_GLOBAL performs dict lookup ad caches the result, and modifying the bytecode at the current pc for LOAD_GLOBAL_CACHED / STORE_GLOBAL_CACHED. The _CACHED opcode loads the cached value always.

Whenever a ValueCell is cached, its user is tracked by ValueCell.dependencyLink. Under a condition that the cached value gets invalid, the invalidated ValueCell is removed from all dependent functions AND we revert all _CACHED opcode to their original opcode.

## Performance Impact

This change decreases the instruction count on Richards by 2% (and this is a gain recovering from a regression experienced by turning on caching for builtins by T45232323 ).

|   | Instruction Count |
| -- | -- |
| Control | 3,358,814,670 |
| Variation | 3,288,234,259 |

## Affected opcodes

The opcode involved with this change are as follows:
- LOAD_GLOBAL: perform global variable lookup and caches it; replace the current opcode for LOAD_GLOBAL_CACHED
- LOAD_GLOBAL_CACHED: cache hit is always expected; load the cached value;
- STORE_GLOBAL: same as LOAD_GLOBAL, but for storing a value; performs cache invalidation if name shadowing happens.
- STORE_GLOBAL_CACHED: cache hit is always expected; store the new value in the cached valuecell.

- DELETE_GLOBAL: Invalidate the cache for the variable to be deleted.
- STORE_NAME: If the global dict (module dict) is overwritten, invalidate the cache entry. Values from typedict are not cached so no special treatment is needed
- DELETE_NAME: same as above.

Note that LOAD_NAME doesn't do caching, and it's intended for simplification.

## Self-modifying code

As noted above, we don't introduce LOAD_GLOBAL_CACHED when a function object is created, but when a cache value is created. This lets us skip a check for cache hit at runtime.

## Follow-up change

Any assignment to `globals()` and `__builtins__` needs to invalidate relevant cache entries, and this will be done by a following diff wrapping dictionaries to track cached values.

**Design Doc**: https://fb.quip.com/2ToaAzSiaYmA
**Umbrella Task**: T44584699

Based on Facebook D15498588