Remove fast globals
**NOT TO BE SUBMITTED ON ITS OWN**

This is an initial change that further changes for implementing global variable caching are going to be stacked. Note that this diff is to be shipped with the future changes to prevent performance degradation.

This change removes fastGlobals from the entire codebase as it puts back a regular non-caching implementation for LOAD_GLOBAL, LOAD_NAME, STORE_GLOBAL, STORE_NAME, and DELETE_GLOBAL alongside with test changes.

Fastglobals are removed from functions/frames. The affected opcodes still assume that global variables are wrapped in ValueCell, however they are explicitly referenced to/from two separate dictionaries, `globals` and `builtins`. Consulting `globals` and `builtins` forced the diff not to remove hidden names ('module') from them, and consistently hiding them from global scopes will be tracked by T39575976.

This diff is *not* attempting to simplify ValueCell usages, and the task is tracked by T44801497.

**Design Doc**: https://fb.quip.com/2ToaAzSiaYmA
**Umbrella Task**: T44584699

Based on Facebook D15437824