Fix global var caching bug of not handling EXTENDED_ARG properly
Summary: The bug was introduced based on a wrong understanding about the order of EXTENDED_ARG and its following opcode. This diff fixes it and addes a test guarding a relevant behavior. Note that T45440363 was filed to properly implement the functionality related to EXTENDED_ARG in the future with bytecode stretching.

Based on Facebook D15682449