Add LOAD_FAST_REAR and STORE_FAST_REAR
The way the stackframe is organized at the moment, means the lowest numbered local is furthest away from the frame pointer:
```
stackframe:
   - local0
   - local1
   - ...
   - local n
...
<-- frame pointer points here
```

This means accessing local `n` requires extra work because we have to compute the localtion of `local0` first and need to load the number of locals. This introduces new opcode `LOAD_FAST_REAR` and `STORE_FAST_REAR` that have `num_locals - x` indexes instead, so we can perform a faster indexing relative to the stack pointer.

Changes the rewriting infrastructure, to allow rewriting any arbitrary bytecode+arg pair into a new bytecode+arg pair and optionally requesting a cache for the new bytecode.

Based on Facebook D15500531