Add argument comment checking to clang-tidy
This checks comments of the form `/*foo=*/` appearing before an
argument for agreement between the name in the comment and the name in
the declaration.  This is especially useful for clarifying the meaning
of a boolean arguments.

Here is an example of a lint failure

```
   Error  (PYROCLANGTIDY) CLANGTIDY
    /home/cshapiro/fbsource/fbcode/pyro/runtime/runtime-test.cpp:65:30:
    warning: argument name 'cache_enabfled' in comment does not match
    parameter name 'cache_enabled' [bugprone-argument-comment]
      Runtime runtime(heap_size, /*cache_enabfled=*/false);
                                 ^~~~~~~~~~~~~~~~~~~
                                 /*cache_enabled=*/
    /home/cshapiro/fbsource/fbcode/pyro/runtime/runtime.h:71:32: note:
    'cache_enabled' declared here
      Runtime(word heap_size, bool cache_enabled);
                                   ^

    >>>        1
```

A description of this check can be found here

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-argument-comment.html

Based on Facebook D15680780