Do not pre-initialize locals
Stop pre-initializing locals. The current plan here is to write a data flow analysis that analysis the (small) set of local accesses that are possibly uninitialized. And then adds `del` instructions for them upfront. This allows us to skip the `memset` in `Thread::openAndLinkFrame()`.

This will now result in an uninitialized `virtualPC` for native functions. This needed some changes:
- Set the `lntotab` field of native code objects to empty bytes (previously we would keep the one of the python declaration, which wasn't too meaningful anyway as the code is C++ code and was not defined at that place the python lnotab indicated).
- Stop the traceback printer from accessing the pc and reporting line numbers when the `lnotab` is empty.

Based on Facebook D15705960