Make all handlers return a bool value
We used to use template magic in `wrapHandler()` in the interpreter, to automatically add a `return false;` in case of void opcode handlers. This removes the template magic in favor of a little more lines.

I expect this to simplify the planned manual assembly changes as we do not need to deal with template function name mangling any longer.

Based on Facebook D15751462