Don't return Unbound values to managed code from BaseException attributes
This is a follow-up to D14447384, to fix the issue referenced in the
description: "if managed code asks for one of these attributes, we'll pass the
UnboundValue back unchanged".

Move `_structseq_{get,set}attr` to `builtins-module.{h,cpp}`, give them
appropriately generic names, and use them in a few descriptors in
`BaseException` to hide `Unbound` values from managed code. Add
`builtins._traceback` for a type check in one of the descriptors (CPython
doesn't expose `traceback` in `builtins`).

Based on Facebook D15561772