Fix UnicodeError attributes and add UnicodeErrorBase
CPython doesn't actually store any attributes on `UnicodeError`, only it's subclasses.

The actual `UnicodeError` type is initialized with no attributes, and only its subclasses use `UnicodeErrorObject`. This diff abstracts out a `UnicodeErrorBase` that each of the subclasses inherit from, and make sure that all usages of a `Unicode.*Error` know to expect a `UnicodeErrorBase` instead of a `UnicodeError`.

This diff makes us match that behavior to allows us to have proper `builtinBase`s.

Based on Facebook D15477705