Implement intrinsic functions
Some builtin functions are simple enough that we can compute them inplace rather than opening up a new stack frame. In other functions, there is a simple special case which can also be handled in place.

Add the symbol of each intrinsic function to its interpreter information. Then, when handling a call, we check whether the function has an intrinsic symbol, and switch on that symbol.

As a prototype, intrinsify `_type` and the builtin type checks.

Based on Facebook D15477286