Use placeholder as cache invalidation signal
`LOAD_GLOBAL` can cache a value either from the module dict or the builtins dict (`__builtins__`). Since a name in the module dict shadows the one in `__builtins__`, a store to the module dict should invalidate cache values from `__builtins__` under the same name.

This change enables this cache invalidation as follows:
- When we cache a value from the builtin dict, store a "placeholder" in the module dict
- When an entry is added to the module dict, check if the existing entry value is a placeholder. If it is, invalidate the cache value derived from the builtins dict under the same entry key.

**Performance Impact**

This change made a minor performance improvement by avoid the second lookup into the builtins dict when a global variable is stored:

| Version | Instruction Count (Richards X 5) |
| -- | -- |
| Original | 2298707380 |
| This Change | 2284569483 |

(Original - This Change) / Original * 100.0 = 0.61%

Based on Facebook D15741141