Add native object tracking
Update `ApiHandle::asInstance` track the native object and avoid memory leaks at runtime destruction time. This modifies ListEntry to have an extra word called `native`. If the `native` word is set, ListEntry is tracking a native object. Otherwise, it is the header of a native object tracked by the GC. Furthermore, given that the native object is not managed by the runtime, we can re-use the `reference_` to point back to the ListEntry.

At native object destruction time, if the object is `native` and has the `reference_` word set, it will untrack and free the ListEntry as well as the object itself.

At runtime destruction time, it will traverse the list of remaining tracked objects. If it has the `native` word set, it will free that native object and also free itself.

Note that `PyObject_Type` was converting non-managed objects into managed ones (to pull the type out). Instead use the `ApiHandle::type` function to pull the type object out without converting a non-managed object into a managed one.

Based on Facebook D15720690