Print exception tracebacks in displayException()
If `__traceback__` is a `str` (which it is in debug builds or when
`PYRO_RECORD_TRACEBACKS` is set), print it in `displayException()`. Also make a
few small tweaks to our traceback formatting to match CPython.

The motivation for this is that we'll now get tracebacks when an exception
escapes to the top level uncaught, like in CPython (see test plan).

Based on Facebook D15835724