Implement a generic dict lookup/insert
This implements a generic dictionary lookup/insertion algorithm. This differs from the internal runtime dictionary lookup/insertion in that it uses the key's `__eq__` to compare both keys.

This uses a similar probing strategy with the difference that it heavily prioritizes the first probe, making that the fast path. If that fails, then it will start probing the rest of the dictionary until it finds a match or determines that the dictionary does not contain an element.

Based on Facebook D15706986