Shrink ObjectFormat from 3 to 1 bit
We only really have two object types: objects that hold raw data and
objects that hold other objects. Take away 2 of the 3 bits, giving them to the
hash code for now (we can always change that in the future).

For data instances, the header count always holds the object's size in bytes
now, which required adjustments to a few size calculations.

Based on Facebook D15836396