Move bytecode + caches from context to frame
Move the `bytecode` and `caches` fields from `Interpreter::Context` to the `Frame` object. This will slightly increase the callframes, but has the advantage that we only need to set the two fields on every call while previously we had to set them on every call and return.

This has a small performance improvement, but the final goal here is to get rid of the whole `Context` structure to simplify the hand assembled interpreter.

Based on Facebook D15834952