Use RawTuple reference to caches
Change the inline cache interfaces to take raw references to the caches tuple. They do not perform any memory allocations, so this is safe. It has no effect on performance.

This change helps in my next change moving the caches field from the interpreter context where it was reference by a handle into the frame object where the getter will only return a raw reference.

Based on Facebook D15834634