Use Interpreter::doFunction() when an opcode handler ends in a call
Using `Interpreter::call()` will recursively enter the interpreter. However if we are at the end of an opcode then we can possibly do a non-recursive (from the C++ point of view) call to the target function by using `doCallFunction()` instead of the various `callXXX()`/`callMethodXXX()` variants.

Add `tailcallMethod1()` and `tailcallMethod2()` convenience functions that can be used instead of `callMethod1()`/`callMethod2()` at the end of an opcode handler.

Replaced calls where possible.

Based on Facebook D15721301