Optimize Frame::setVirtualPC() / virtualPC()
Summary: `setVirtualPC()` / `virtualPC()` is called at least once for every operation and is therefore very performance sensitive. Given that the PC is naturally an even number with the lowest bit cleared allows us to directly store its value instead of shifting it in place and back when used as a proper `SmallInt`.

Based on Facebook D15876708