Remove interpreter context
Summary: There is nothing left in the interpreter context except a pointer to the current thread and a pointer to the current frame. Thread has a pointer to the current frame as well. So we can just remove the interpreter context now and have all the users take a thread instead. This saves some indirections where we previously went from context to thread, and we only have to update the current frame in one place instead of two.

Based on Facebook D15889000