Unwind outside opcode handler, remove entry_frame from context
This changes the opcode handler to return an enum value.
The nice thing about this is that we have a slightly easier time implementing then unwinding/return logic outside the opcode handlers, which by nature cannot make the whole interpreter return.

This allows us to move the `entry_frame` field out of the `Interpreter::context` struct and make it a local variable in `execute()` only.

This should be about as-fast as before for the `Continue::NEXT` case. The compiler also should typically apply jump-threading to the `return Continue::XXX; ... if (cont == Continue::XXX) { ... }` cases and after inlining replace the return with a direct jump into the later then-block.

It may be slightly slower where the compiler cannot if-convert but that should only affect slow/rare cases and on the bright side avoids alot of code bloat by only having the `unwind()` call in one place.

Based on Facebook D15888671