Split finishCurrentGenerator() into fast and slow part
Outline the slow part of `finishCurrentGenerator()` into the new `markGeneratorFinished()` function and mark the fast-path `HANDLER_INLINE`.

While on it also pull out the `finishGenerator()` call from `popFrame()` so we can merge two independent ones in `handleReturn()` (and because I find it somewhat unintuitive to have this happen as part of `popFrame()` anyway).

Based on Facebook D15904524