Add blob storage to send binary from client to server
This enables the client to send a binary to the server by using a blob storage service.

The client will get the path to the binary it wants to benchmark. This will be sent to Everstore which will return a handle. This handle will then be passed to the server which will use it to download and make an executable binary out of that blob.

This uses the newly created `EVERSTORE_PYRO_BINARIES` FbType which have a retention rate of 1 day: https://our.intern.facebook.com/intern/everstore/customer/2136/.

Based on Facebook D15995323