Remove non frozen imports from benchmark runner
As a standalone binary, pyro can only import frozen libraries. Fortunately, the only non frozen library needed for the benchmark suite is `os` to find the benchmark file paths.

This modified `run.py` to obtain the absolute paths of the benchmarks and passes that as an argument to `_benchmark_runner.py`. This avoids having to import a non frozen library.

Based on Facebook D16022237