Enable client and server to run in the localhost
Summary: Add a localhost flag for both the client and server to both in the same machine. This also fixed a couple of bugs such as the import differences between `buck run` and a normal `python` binary. This also adds a `LOCALHOST_PORT` to an arbitrary high number value (34193). If this becomes an issue in the future, we can add it as an argument to the user.

Based on Facebook D16021522