Add a REPL
Introduce a REPL to Pyro. The REPL can be started in the standard way, by
running:

```
maple% ./python
>>> 1 + 2
3
>>> print("I'm a real python!")
I'm a real python!
>>>
maple%
```

This diff introduces the following API functions:
* `PyRun_InteractiveLoopFlags`
  * Drive the REPL, print prompts, etc
* `PyRun_InteractiveOneObjectEx`
  * Parse, compile, load, and execute a module. This is an auxiliary function
    for `PyRun_InteractiveOneObject` that does not print the error on failure.
* `runInteractive`
  * A compiler interface function for `PyRun_AnyFileExFlags`
* `PyRun_AnyFileExFlags`
  * Run a file either interactively or otherwise. This is wholesale copied from
    CPython.

This diff also:
* Fixes a small memory leak in `PyParser_ASTFromFileObject`.
* Does some minor renaming in `pythonrun.cpp` to please the linter.

This REPL is somewhat hamstrung by the fact that we don't have a full IO
implementation up and running, but it shouldn't need to change once we do.

Based on Facebook D15967221