Add benchmark mains
Summary: The new benchmark runner directly calls `run()` for the benchmarks (a variable amount of times). For external tools such as profilers it is useful to run without the runner for a fixed amount of iterations. Add some main code back to the benchmark that does that.

Based on Facebook D15989446