Invalidate only type-matching entries in cache
Currently we empties all entries in a cache for a PC (even for polymorphic caches) if any of the contained values gets invalidated ignoring other values in the cache.

The problem invalidates an invariant on `ValueCell::dependencyLink` that it only needs to track dependents which are currently caching it by illegally deleting too many caches.

This fixes the issue by deleting only affected caches whose type matches the invalidated type, and compresses the cache line for optimizing future calls of `icLookup`.

Based on Facebook D16113445