Implement type.__subclasses__()
Implement the method to return a new list of the type's direct subclasses.

Store the subclasses internally as a dict of weakrefs, using the LayoutId as a unique identifier.

For now, we only modify subclasses on type creation. Once `__bases__` is fully implemented with a setter, it will be possible to modify subclasses further.

Based on Facebook D16166843