Add x64 Assembler class from Dart VM
This is heavily adapted from:
- https://github.com/dart-lang/sdk/blob/1e24fe7d699a1c36be142afa21859c6a9c82d035/runtime/vm/constants_x64.h
- https://github.com/dart-lang/sdk/tree/1e24fe7d699a1c36be142afa21859c6a9c82d035/runtime/vm/compiler/assembler

The files have been modified to fit our style and needs, with Dart-specific
bits stripped out. This was mostly: 1) lowercasing the first letter of method
names, and 2) converting `assert(foo)` to `DCHECK(foo, "assert()")`.

I've imported the Dart LICENSE, AUTHORS, and PATENTS files as well with a -Dart
suffix, and modified the copyright headers to include Facebook (for our
modifications).

Based on Facebook D16019854