Add assembly interpreter
This adds an alternate implementation of our interpreter, generated at
process startup. See `interpreter-gen-x64.cpp` for design and implementation
details. I've implemented enough opcodes to make this a small win (see test
plan for numbers), so it's enabled by default, with an option to disable it.

Based on Facebook D16019855