Run benchmark tools as linux commands
Refactor `run.py` to run the registered tools in `_tools.py` as linux commands. This integrates `_benchmark_runner.py` into `run.py` and modifies `_tools.py` so that they are expected to run a linux command.

This also moves the dynamic iteration run of the benchmark to be run as an external python file in `_time_tool.py`

Based on Facebook D16194877