Add platform007.cmake pre-load script
Add cmake script that to be used with the `-C` flag of `cmake` that configures a build suitable for deployment on facebook platform007 machines.

This is not meant to be the final word on decisions like clang vs gcc, what optimization flags to use,
using libraries from `/usr/local/fbcode/platform007/lib/` vs `/lib64` etc. We can tweak those settings
in follow-ups.

Based on Facebook D16182932