Allow iteration over Sequence classes
CPython allows for iteration over anything that implements the `Sequence` protocol. This ends up just meaning anything that implements a `__getitem__` method.

This diff adds CPython's `iterator` class (which is the type the object returned by calling `iter` on a `Sequence`), and makes `iter` and `doGetIter` instantiate and return it when passed in a `Sequence` object.

It follows the same format as the other `iterator` types defined in `builtins.py`.

Based on Facebook D16012886