Switch ByteArray to use MutableBytes and remove LargeBytes::byteAtPut
`ByteArray` is currently backed by either en empty `Bytes` object or a `LargeBytes`. This causes unnecessary dispatching overhead and forces us to break our invariant that `Bytes` should be immutable. We can fix this by switching it to be backed by `MutableBytes` instead.

This diff makes `MutableBytes` be a subclass of `LargeBytes`, and changes any assumption that `ByteArray` stores a `Bytes` object that isn't compatible with `MutableBytes` as well.

Since this is the last thing that requires `LargeBytes` to be mutable, the diff also makes `LargeBytes` immutable

Based on Facebook D16090581