Expand state avilable during assembly interpreter generation
I'm playing around with a few more ideas for speeding up the assembly
interpreter, and they all need to share more state between the different emit
functions. This diff prepares for that by passing around a state structure
rather than just an `Assembler*`. I probably should've structured things like
this from the beginning, but better late than never.

For now, the only added state is the current `Bytecode`, used by
`emitGenericHandler()` to avoid some duplication in the callers, and a string
description of the opcode (because the 3 pseduo-handlers for control flow
operations don't have corresponding `Bytecode`s). The latter makes for a much
nicer failure message if we overflow a handler.

Based on Facebook D16267069