Avoid accessing dependency list while updating it
Currently `icInvalidateCachesForTypeAttr` iterates over a linked list and calls `icDeleteCacheForTypeAttrInDependent` to modify the list being iterated.

This is error prone since `icDeleteCacheForTypeAttrInDependent` may modify the currently processed node's contents (e.g., prev, next pointers) or delete anything not unprocessed in the list. This is currently not happening but it may happen with a change to the function.

This change makes sure that `icInvalidateCachesForTypeAttr` iterates over all dependents by capturing the next pointer before modifying the current node.

Based on Facebook D16171200