Pull in _sre
Since `_Py_ascii_whitespace` is defined in `unicodeobject.c`, we cannot pull in CPython's implementation, so we just define it ourselves.

Since our `PyUnicode_DATA` function currently just returns a `utf8` buffer, it is possible to properly index into ASCII characters if we also define `PyUnicode_KIND` to be `1BYTE_KIND`.

This imports a few files from `third-party/cpython` to allow us to take advantage of `_sre` functions that we can pull in whole-sale.

Based on Facebook D16183147