Enable metaclass use in enum module
Most of the enum module was disabled before because we lacked the necessary metaclass support. Now we can enable it!

Having enum working means `RegexFlags.__members__` is available. This enables a better workaround for some code in the `re` module.

Just had to insert a couple TODOs + workarounds because `<sometype>.__dict__` currently gives you a dictionary with valuecells that the code did not expect.

Based on Facebook D16442449