Fix _isinstance_type nargs bug
In diff review, I changed `_isinstance_type` to take the type of the first argument as a second argument, but I missed a callsite in `type.__instancecheck__`.

Now that we can execute this call through `isinstance` with a custom metaclass, this error pops up when attempting to `import re`, which uses enum.

Based on Facebook D16442994