Import re
There are a few old commented out classes and variables in `re` and `sre_constants` that are remnants of our inability to subclass ints. Since we can do that now, we can uncomment them out. Same as in D15018310.

This allows us to now `import re`.

Based on Facebook D16421033