Add support for more than two interpreters in benchmark client
This modifies the client to allow it to run more than one interpreter. Given that the output table will now increase, this also modifies the formatting to return a markdown table to make it easier to read.
This also adds more logging on the client for better debuggability

Based on Facebook D16447578