Don't allocate a tuple for slice operations
Calling `_foo_getslice(self, *key.indices(...))` is slow because we allocate
and then re-explode a tuple onto the stack using `CALL_FUNCTION_EX`. Calling
`start, stop, step = key.indices(...)` is slow because we allocate a tuple (but
better because we do not re-explode the tuple). This diff makes slicing
operations faster by avoiding tuple allocation when processing a `slice`.

This diff adds a `None` case to `_slice_index` that returns immediately. Since
there is only one place that uses this in our runtime, `sliceIndex` (in slice-builtins),
and the users of sliceIndex separately account for `None`, this change has
minimal impact.

It introduces `_slice_step`, `_slice_start`, and `_slice_stop` functions and
then re-uses them in `slice.indices`.

Based on Facebook D16290441