Simplify the MutableBytes interface
`MutableBytes` had two methods whose names were qualified with
`Bytes` but there was no agreement on what type that was.  One
method accepted a `View<bytes>` but the other accepted a `Bytes`
instance.  This change cleans up the interface for consistency
with other types by removing a method and renaming the remaining
method.

* Re-implement `isMutableBytesEqualsBytes` to delegate to
  `isBytesEqualsBytes`, the only use of `compareWithBytes`.

* Remove `compareWithBytes` now that it's dead.

* Shorten `replaceFromWithBytes` to `replaceFromWith` as the
  extra qualification is not significant

Based on Facebook D16461525