Implement range.__len__
Implement the builtin method, sharing code with `longrange_iterator.__length_hint__`.

For performance reasons, this requires changes to the values that we store on `RawRange`. The subscript operation requires calculating the length, so repeated calls to `__getitem__` would be a serious issue without storing the length of the range the first time it is calculated.

Based on Facebook D16465694